[IMP] <pro>DataSet.Field: Optimize some problems of chain binding.
